# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateFileResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      file_hash:
        type: string
      file_name:
        type: string
      metadata:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
...
  api.FileItemResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      file_hash:
        type: string
      file_name:
        type: string
      file_url:
        type: string
      metadata:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
  api.FileResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      file_hash:
        type: string
      file_name:
        type: string
      metadata:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
...
  api.UpdateFileRequest:
    properties:
      metadata:
        type: string
    type: object
...
info:
  contact: {}
  description: Менеджер паролей GophKeeper
  title: GophKeeper API
  version: "1.0"
paths:
...
  /api/file:
    get:
      description: Возвращает список файлов, загруженных текущим пользователем
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.FileItemResponse'
            type: array
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список файлов пользователя
      tags:
      - file
  /api/file/{file_hash}/{file_name}:
    delete:
      description: Удаляет файл пользователя из хранилища вместе с его метаданными
      parameters:
      - description: Хэш файла
        in: path
        name: file_hash
        required: true
        type: string
      - description: Имя файла
        in: path
        name: file_name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Файл не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление файла
      tags:
      - file
    get:
      description: Возвращает содержимое файла пользователя по хэшу и имени файла
      parameters:
      - description: Хэш файла
        in: path
        name: file_hash
        required: true
        type: string
      - description: Имя файла
        in: path
        name: file_name
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Содержимое файла
          schema:
            type: file
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "404":
          description: Файл не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение файла
      tags:
      - file
    put:
      consumes:
      - multipart/form-data
      description: |-
        Перезаписывает содержимое файла пользователя. Хэш, имя и метаданные остаются прежними,
        обновляются контрольная сумма, размер и время загрузки
      parameters:
      - description: Хэш файла
        in: path
        name: file_hash
        required: true
        type: string
      - description: Имя файла
        in: path
        name: file_name
        required: true
        type: string
      - description: Новое содержимое файла
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FileResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Файл не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновление файла
      tags:
      - file
  /api/file/{file_hash}/{file_name}/metadata:
    put:
      consumes:
      - application/json
      description: Меняет только метаданные файла, содержимое в хранилище остаётся
        прежним
      parameters:
      - description: Хэш файла
        in: path
        name: file_hash
        required: true
        type: string
      - description: Имя файла
        in: path
        name: file_name
        required: true
        type: string
      - description: Новые метаданные
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.FileResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Файл не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновление метаданных файла
      tags:
      - file
  /api/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл в S3-хранилище и сохраняет его метаданные для текущего
        пользователя
      parameters:
      - description: Загружаемый файл
        in: formData
        name: file
        required: true
        type: file
      - description: Произвольные метаданные
        in: formData
        name: metadata
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateFileResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "409":
          description: Файл с таким содержимым уже загружен
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Загрузка файла
      tags:
      - file
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"